feat: remove the temp accounts for deloying the ref utxos holdig the scripts
