+++ cardano/gateway/.env.example
@@ -12,7 +12,7 @@ OGMIOS_ENDPOINT=http://127.0.0.1:1337
 CARDANO_CHAIN_HOST=127.0.0.1
 CARDANO_CHAIN_PORT=3001
 CARDANO_NETWORK_MAGIC=42
-CARDANO_EPOCH_NONCE_GENESIS="3c30e01c4d2be5cc21e14a88542929e5f9c2cfb7af6d94a746d7d61b6049dc62"
+CARDANO_EPOCH_NONCE_GENESIS="3c1195a3fb825b2d2a6ba4e67008288ab7933a6db263a28eb9fd19b3ca05212f"
 
 HANDLER_JSON_PATH=/usr/src/app/cardano/offchain/deployments/handler.json
