+++ cardano/offchain/src/utils.ts
@@ -4,30 +4,17 @@ import {
   validatorToAddress,
   Address,
   applyParamsToScript,
-  Blockfrost,
   Data,
-  Emulator,
   Exact,
   fromHex,
   fromText,
-  Kupmios,
-  Lucid,
-  PROTOCOL_PARAMETERS_DEFAULT,
-  Provider,
   Script,
   ScriptHash,
-  SLOT_CONFIG_NETWORK,
   toHex,
   TxBuilder,
   UTxO,
   LucidEvolution,
 } from "@lucid-evolution/lucid";
-import {
-  BLOCKFROST_ENV,
-  EMULATOR_ENV,
-  KUPMIOS_ENV,
-  LOCAL_ENV,
-} from "./constants.ts";
 import { AuthToken } from "../lucid-types/ibc/auth/AuthToken.ts";
 import { OutputReference } from "../lucid-types/cardano/transaction/OutputReference.ts";
 import { crypto } from "@std/crypto";
@@ -78,7 +65,7 @@ export const submitTx = async (
   const txHash = await signedTx.submit();
   console.log("Submitting tx [", txName, "]: tx hash is", txHash);
   console.log("Submitting tx [", txName, "]: waiting for adoption ...");
-  await lucid.awaitTx(txHash, 2000);
+  await lucid.awaitTx(txHash, 1000);
   console.log("Submitting tx [", txName, "]: done");
   return txHash;
 };
@@ -102,61 +89,6 @@ export const formatTimestamp = (timestampInMilliseconds: number): string => {
 export type Signer = {
   sk: string;
   address: string;
-};
-
-export const setUp = async (
-  mode: string
-): Promise<{ lucid: LucidEvolution; signer: Signer; provider: Provider }> => {
-  const signer = {
-    sk: "ed25519_sk1rvgjxs8sddhl46uqtv862s53vu4jf6lnk63rcn7f0qwzyq85wnlqgrsx42",
-    address: "addr_test1vz8nzrmel9mmmu97lm06uvm55cj7vny6dxjqc0y0efs8mtqsd8r5m",
-  };
-  let provider: Provider;
-  let lucid: LucidEvolution;
-  if (mode == EMULATOR_ENV) {
-    console.log("Deploy in Emulator env");
-    provider = new Emulator(
-      [
-        { address: signer.address, assets: { lovelace: 3000000000000n }, privateKey: signer.sk, seedPhrase: '' },
-        { address: signer.address, assets: { lovelace: 3000000000000n }, privateKey: signer.sk, seedPhrase: '' },
-      ],
-      { ...PROTOCOL_PARAMETERS_DEFAULT, maxTxSize: 900000 }
-    );
-    lucid = await Lucid(provider, "Preview");
-  } else if (mode == KUPMIOS_ENV) {
-    const kupo = "http://192.168.10.136:1442";
-    const ogmios = "http://192.168.10.136:1337";
-    console.log("Deploy in Kupmios", kupo, ogmios);
-    provider = new Kupmios(kupo, ogmios);
-    const chainZeroTime = await querySystemStart(ogmios);
-    SLOT_CONFIG_NETWORK.Preview.zeroTime = chainZeroTime;
-    lucid = await Lucid(provider, "Preview");
-  } else if (mode == LOCAL_ENV) {
-    const kupo = "http://localhost:1442";
-    const ogmios = "http://localhost:1337";
-    console.log("Deploy in local", kupo, ogmios);
-    provider = new Kupmios(kupo, ogmios);
-
-    const chainZeroTime = await querySystemStart(ogmios);
-    SLOT_CONFIG_NETWORK.Preview.zeroTime = chainZeroTime;
-    lucid = await Lucid(provider, "Custom");
-  } else if (mode == BLOCKFROST_ENV) {
-    provider = new Blockfrost(
-      "https://cardano-preview.blockfrost.io/api/v0",
-      "preview2fjKEg2Zh687WPUwB8eljT2Mz2q045GC"
-    );
-    lucid = await Lucid(provider, "Preview");
-  } else {
-    throw new Error("Invalid provider type");
-  }
-
-  lucid.selectWallet.fromPrivateKey(signer.sk);
-
-  return {
-    lucid,
-    signer,
-    provider,
-  };
 };
 
 export const generateTokenName = async (
@@ -212,11 +144,13 @@ export const querySystemStart = async (ogmiosUrl: string) => {
 
   client.addEventListener('open', () => console.log('WebSocket connection opened.'));
   client.addEventListener('close', (event) => {
-    console.log('WebSocket connection closed.', {
-      code: event.code,
-      reason: event.reason,
-      wasClean: event.wasClean,
-    });
+    if (!event.wasClean) {
+      console.log('WebSocket connection closed.', {
+        code: event.code,
+        reason: event.reason,
+        wasClean: event.wasClean,
+      });
+    }
   });
   client.addEventListener('error', (err) => console.log('WebSocket error:', err));
 
@@ -401,3 +335,51 @@ export const getNonceOutRef = async (
 
   return [NONCE_UTXO, outputReference];
 };
+
+type Validator =
+  | "spendHandler"
+  | "mintClient"
+  | "spendClient"
+  | "mintConnection"
+  | "spendConnection"
+  | "mintChannel"
+  | "spendChannel"
+  | "mintPort"
+  | "mintIdentifier"
+  | "spendTransferModule"
+  | "mintVoucher"
+  | "verifyProof";
+
+type Module = "handler" | "transfer";
+
+type Tokens = "mock";
+
+export type DeploymentTemplate = {
+  validators: Record<
+    Validator,
+    {
+      title: string;
+      script: string;
+      scriptHash: string;
+      address: string;
+      refUtxo: UTxO;
+      refValidator?: Record<
+        string,
+        { script: string; scriptHash: string; refUtxo: UTxO }
+      >;
+    }
+  >;
+  handlerAuthToken: {
+    policyId: string;
+    name: string;
+  };
+  modules: Record<
+    Module,
+    {
+      identifier: string;
+      address: string;
+    }
+  >;
+  tokens: Record<Tokens, string>;
+};
+
